requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
RPi.GPIO==0.7.1 